+++ package.json
@@ -1,0 +1,54 @@
+{
+  "name": "rc-timeline-carousel",
+  "version": "1.0.4",
+  "description": "UI Component in the SNS form of Timeline Feed Slider for Developer with REACT",
+  "keywords": [
+    "react",
+    "react-component",
+    "react-slider",
+    "sns-carousel",
+    "horizontal",
+    "horizontal-slider",
+    "carousel",
+    "timeline",
+    "image",
+    "video",
+    "feed",
+    "feed-form"
+  ],
+  "main": "build/index.js",
+  "style": "./css/Timeline.css",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:kimyongyook/react-timeline-carousel.git"
+  },
+  "bugs": {
+    "url": "http://github.com/kimyongyook/react-timeline-carousel/issues"
+  },
+  "peerDependencies": {
+    "react": "^16.9.0"
+  },
+  "scripts": {
+    "start": "webpack --watch",
+    "build": "webpack",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "react": "^16.9.0",
+    "webpack": "^4.39.2"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "babel-preset-env": "^1.7.0",
+    "css-loader": "^3.2.0",
+    "file-loader": "^4.2.0",
+    "style-loader": "^1.0.0",
+    "webpack-cli": "^3.3.6"
+  },
+  "author": "kimyongyook",
+  "license": "MIT"
+}
